Factor out column related functions
